+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Add new setting `:completion :analysis-type` to choose between `:fast-but-stale`(default) or `:slow-but-accurate`, this should define whether completion should wait for changes that may still happening, this by default reverts the behavior introduced after #1425. #1487
   - Fix `textDocument/hover` issue when doc metadata isn't a string literal.
   - Follow references to other vars in doc metadata for use in `textDocument/hover`.
+  - Support `:arglists` meta when finding docs of functions via hover feature. #1504
 
 ## 2023.01.26-11.08.16